# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UserController_getMyProfile",
        "summary": "",
        "description": "Get profile user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponseDto" } } }
          }
        },
        "tags": ["Users"],
        "security": [{ "bearer": [] }]
      }
    },
    "/users": {
      "post": {
        "operationId": "UserController_create",
        "summary": "",
        "description": "Create new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateUserRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponseDto" } } }
          },
          "400": { "description": "Incorrect data." },
          "500": { "description": "Internal server error" }
        },
        "tags": ["Users"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "UserController_getAll",
        "summary": "",
        "description": "Get all user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/UserResponseDto" } }
              }
            }
          }
        },
        "tags": ["Users"],
        "security": [{ "bearer": [] }]
      }
    },
    "/users/{id}": {
      "put": {
        "operationId": "UserController_update",
        "summary": "",
        "description": "Update user",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "multipart/form-data": { "schema": { "$ref": "#/components/schemas/UpdateUserRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponseDto" } } }
          },
          "404": { "description": "There isn't any user with id" },
          "500": { "description": "Internal server error" }
        },
        "tags": ["Users"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "UserController_findById",
        "summary": "",
        "description": "Find by ID user",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponseDto" } } }
          },
          "500": { "description": "Internal server error" }
        },
        "tags": ["Users"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "UserController_delete",
        "summary": "",
        "description": "Delete user",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" }, "404": { "description": "There isn't any user with id" } },
        "tags": ["Users"],
        "security": [{ "bearer": [] }]
      }
    },
    "/users/byIds": {
      "delete": {
        "operationId": "UserController_deleteByIds",
        "summary": "",
        "description": "Delete users",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteByIdsRequestDto" } } }
        },
        "responses": {
          "200": {
            "description": "Undeleted ids",
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/" } } }
            }
          },
          "404": { "description": "There isn't any user with id" }
        },
        "tags": ["Users"],
        "security": [{ "bearer": [] }]
      }
    },
    "/users/search": {
      "get": {
        "operationId": "UserController_search",
        "summary": "",
        "description": "Search users",
        "parameters": [
          { "name": "name (first or last)", "required": true, "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/UserResponseDto" } }
              }
            }
          }
        },
        "tags": ["Users"],
        "security": [{ "bearer": [] }]
      }
    },
    "/users/save-last-session": {
      "post": {
        "operationId": "UserController_saveLastSessionDateTime",
        "summary": "",
        "description": "Update session date and time",
        "parameters": [],
        "responses": { "200": { "description": "" }, "500": { "description": "Internal server error" } },
        "tags": ["Users"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/access/user/{id}": {
      "put": {
        "operationId": "AccessController_updateUser",
        "summary": "",
        "description": "Update user to admin",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "" } },
        "tags": ["Access"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": { "title": "BACKEND", "description": "REST API", "version": "1.0.0", "contact": {} },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "DepartmentResponseDto": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "name": { "type": "string" } },
        "required": ["id", "name"]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "role": { "type": "string" },
          "skills": { "type": "string" },
          "lastSessionDateTime": { "type": "string" },
          "avatarPath": { "type": "string" },
          "phone": { "type": "string" },
          "department": { "$ref": "#/components/schemas/DepartmentResponseDto" }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt",
          "firstName",
          "lastName",
          "role",
          "skills",
          "lastSessionDateTime",
          "avatarPath",
          "phone",
          "department"
        ]
      },
      "CreateUserRequestDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "description": "User email", "example": "email@gmail.com" },
          "firstName": { "type": "string", "description": "User FirstName", "example": "Alex" },
          "lastName": { "type": "string", "description": "User LastName", "example": "Grabovsky" },
          "phone": { "type": "string", "description": "User phone number", "example": "+380636479067" }
        },
        "required": ["email", "firstName", "lastName", "phone"]
      },
      "UpdateUserRequestDto": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string", "description": "User  FirstName", "example": "Alex" },
          "lastName": { "type": "string", "description": "User LastName", "example": "Bin" },
          "email": { "type": "string", "description": "User email", "example": "email@gmail.com" },
          "skills": {
            "example": [
              "UI/UX Design (WEB/Mobile)",
              "Wireframing",
              "Prototyping",
              "User Research",
              "Interaction design",
              "Responsive Web Design",
              "CJM"
            ],
            "description": "Array of skills",
            "type": "array",
            "items": { "type": "string" }
          },
          "phone": { "type": "string", "description": "User phone number", "example": "+380636479067" },
          "file": { "type": "string", "format": "binary" },
          "departmentId": {
            "type": "string",
            "description": "Department ID",
            "example": "7fd524bc-ebaa-4a34-8202-6fa0cb1e3878"
          }
        },
        "required": ["firstName", "lastName", "email", "departmentId"]
      },
      "DeleteByIdsRequestDto": {
        "type": "object",
        "properties": {
          "ids": {
            "example": ["9322c384-fd8e-4a13-80cd-1cbd1ef95ba8", "986dcaf4-c1ea-4218-b6b4-e4fd95a3c28e"],
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["ids"]
      },
...
    }
  }
}